requirements.txt
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
click==8.3.1
fastapi==0.128.0
h11==0.16.0
httptools==0.7.1
idna==3.11
inflect==7.5.0
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==10.8.0
psycopg==3.3.2
psycopg2-binary==2.9.10
pydantic==2.12.5
python-jose[cryptography]==3.3.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sqlacodegen==3.2.0
SQLAlchemy==2.0.35
starlette==0.50.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
